requirements.txt
argcomplete # script's args
websocket-client  # HA WebSocket API (discover devices from HA registry)
pyyaml # config files
types-pyyaml  # typing for pyyaml
webcolors  # get names for RGB colors

# Network Communication
httpx[http2]  # HTTP client for Python with HTTP2 support # faster than requests
urllib3  # use to disable warnings

# Hue communication
# https://github.com/hrdasdominik/hue-entertainment-pykit
# dependencies: https://github.com/hrdasdominik/hue-entertainment-pykit/blob/main/pyproject.toml
# hue-entertainment-pykit requires Python<=3.12
hue-entertainment-pykit

# python-mbedtls==2.10.1  # the newest version # https://pypi.org/project/python-mbedtls/2.10.1/
